feat: add fully working Counter struct
An AatTree is maintained within a Counter to handle all the word counting.

Find and resolve some un-freed pointer problems in aat.c. The major issue are: 1) pointers are not freed when a method is returned early; 2) key_process called twice and the pointer made in the middle gets lost and is never freed.