fix: solve the inifinite loop that is blocking the multi process from completing
The problem is caused by a NodeStack declared on stack memory, which gets out of scope when the
iterator_make() method finishes, but the NodeStack is then used as a heap allocated struct.
The problem is solved by changing NodeStack to heap allocated, and free() it in iterator_free().

Also refactor multi process main to display the final word counts. A copy of the output is stored
in output.txt.